services/api/requirements.txt
fastapi==0.118.*
uvicorn[standard]==0.30.*
SQLAlchemy==2.0.*
psycopg[binary]==3.2.*
alembic==1.13.*
pydantic==2.*
pydantic-settings==2.*
python-multipart==0.0.*
bcrypt==4.0.1
passlib[bcrypt]==1.7.*
email-validator==2.*
python-dotenv==1.*
PyJWT==2.9.*
jinja2==3.1.*
